# specs/001-supplier-service/contracts/openapi-v1.yaml
openapi: 3.1.0
info:
  title: MALIEV Supplier Service API
  description: |
    The Supplier Service is the authoritative source of supplier information within the MALIEV microservices ecosystem.
    It provides APIs for managing supplier records, contacts, certifications, performance evaluations, and audit trails.
  version: 1.0.0
  contact:
    name: MALIEV Development Team
    email: dev@maliev.com
servers:
  - url: http://localhost:8080
    description: Local development
  - url: https://api.maliev.com
    description: Production
...
paths:
  /suppliers/v1/suppliers:
    get:
      tags: [Suppliers]
      operationId: listSuppliers
      summary: List suppliers with filtering and pagination
      description: Returns a paginated list of suppliers with optional filters
      parameters:
        - $ref: '#/components/parameters/PageNumber'
        - $ref: '#/components/parameters/PageSize'
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/SupplierStatus'
        - name: categoryId
          in: query
          schema:
            type: string
            format: uuid
        - name: capability
          in: query
          schema:
            type: string
        - name: search
          in: query
          description: Search by company name
          schema:
            type: string
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [name, createdAt, rating]
            default: name
        - name: sortOrder
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: asc
      responses:
        '200':
          description: Paginated list of suppliers
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SupplierListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
      security:
        - bearerAuth: []

    post:
      tags: [Suppliers]
      operationId: createSupplier
      summary: Register a new supplier
      description: Creates a new supplier with initial status "PendingApproval"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSupplierRequest'
      responses:
        '201':
          description: Supplier created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SupplierResponse'
          headers:
            Location:
              schema:
                type: string
              description: URL of created supplier
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: Duplicate tax ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - bearerAuth: []
  /suppliers/v1/suppliers/{id}:
    parameters:
      - $ref: '#/components/parameters/SupplierId'

    get:
      tags: [Suppliers]
      operationId: getSupplier
      summary: Get supplier details
      description: Returns complete supplier profile including contacts, certifications, and performance summary
      responses:
        '200':
          description: Supplier details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SupplierDetailResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
      security:
        - bearerAuth: []

    put:
      tags: [Suppliers]
      operationId: updateSupplier
      summary: Update supplier information
      description: Updates supplier fields. Supports partial updates.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSupplierRequest'
      responses:
        '200':
          description: Supplier updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SupplierResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Concurrency conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - bearerAuth: []

    delete:
      tags: [Suppliers]
      operationId: deleteSupplier
      summary: Delete a supplier
      description: |
        Deletes a supplier if not referenced by other services.
        Performs synchronous checks with Purchase Order, Invoice, and Stock services.
      responses:
        '204':
          description: Supplier deleted successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Supplier has active references
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DependencyErrorResponse'
        '503':
          description: Dependency service unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - bearerAuth: []
  /suppliers/v1/suppliers/{id}/status:
    parameters:
      - $ref: '#/components/parameters/SupplierId'

    patch:
      tags: [Suppliers]
      operationId: updateSupplierStatus
      summary: Update supplier status
      description: Changes supplier operational status (Active, Suspended, Inactive)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateStatusRequest'
      responses:
        '200':
          description: Status updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SupplierResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
      security:
        - bearerAuth: []
...
  /suppliers/v1/suppliers/{id}/validate:
    parameters:
      - $ref: '#/components/parameters/SupplierId'

    get:
      tags: [Integration]
      operationId: validateSupplier
      summary: Validate supplier exists and is queryable
      description: Used by other services to validate supplier references
      responses:
        '200':
          description: Supplier is valid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SupplierValidationResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
      security:
        - bearerAuth: []
  /suppliers/v1/suppliers/{id}/eligibility:
    parameters:
      - $ref: '#/components/parameters/SupplierId'

    get:
      tags: [Integration]
      operationId: checkEligibility
      summary: Check supplier eligibility for purchase orders
      description: Returns eligibility status based on active status and valid certifications
      responses:
        '200':
          description: Eligibility check result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SupplierEligibilityResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
      security:
        - bearerAuth: []

  # Material Categories (Lookup)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    SupplierId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Supplier unique identifier
    PageNumber:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: pageSize
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorResponse'
  schemas:
...
    SupplierStatus:
      type: string
      enum: [PendingApproval, Active, Suspended, Inactive]
    OnboardingStage:
      type: string
      enum: [PendingApproval, DocumentationReview, FinalApproval, Active]
    CertificationType:
      type: string
      enum: [TaxForm, BusinessLicense, RegulatoryCompliance, QualityCertification, InsuranceCertificate, Other]
...
    CreateSupplierRequest:
      type: object
      required: [companyName, taxId, address, city, country]
      properties:
        companyName:
          type: string
          maxLength: 200
        taxId:
          type: string
          maxLength: 50
        address:
          type: string
          maxLength: 500
        city:
          type: string
          maxLength: 100
        country:
          type: string
          maxLength: 100
        postalCode:
          type: string
          maxLength: 20
        materialCategoryIds:
          type: array
          items:
            type: string
            format: uuid
        capabilities:
          type: array
          items:
            type: string
        primaryContact:
          $ref: '#/components/schemas/CreateContactRequest'
    UpdateSupplierRequest:
      type: object
      required: [rowVersion]
      properties:
        companyName:
          type: string
          maxLength: 200
        address:
          type: string
          maxLength: 500
        city:
          type: string
          maxLength: 100
        country:
          type: string
          maxLength: 100
        postalCode:
          type: string
          maxLength: 20
        materialCategoryIds:
          type: array
          items:
            type: string
            format: uuid
        rowVersion:
          type: string
          format: byte
          description: Base64-encoded row version for concurrency
    UpdateStatusRequest:
      type: object
      required: [status, rowVersion]
      properties:
        status:
          $ref: '#/components/schemas/SupplierStatus'
        reason:
          type: string
          maxLength: 500
        rowVersion:
          type: string
          format: byte
...
    CreateContactRequest:
      type: object
      required: [name, email]
      properties:
        name:
          type: string
          maxLength: 200
        role:
          type: string
          maxLength: 100
        email:
          type: string
          format: email
          maxLength: 255
        phone:
          type: string
          maxLength: 50
        isPrimary:
          type: boolean
          default: false
...
    SupplierResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        companyName:
          type: string
        taxId:
          type: string
        address:
          type: string
        city:
          type: string
        country:
          type: string
        postalCode:
          type: string
        status:
          $ref: '#/components/schemas/SupplierStatus'
        onboardingStage:
          $ref: '#/components/schemas/OnboardingStage'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        rowVersion:
          type: string
          format: byte
    SupplierListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/SupplierResponse'
        totalCount:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer
        totalPages:
          type: integer
    SupplierDetailResponse:
      allOf:
        - $ref: '#/components/schemas/SupplierResponse'
        - type: object
          properties:
            contacts:
              type: array
              items:
                $ref: '#/components/schemas/ContactResponse'
            materialCategories:
              type: array
              items:
                $ref: '#/components/schemas/MaterialCategoryResponse'
            capabilities:
              type: array
              items:
                $ref: '#/components/schemas/CapabilityResponse'
            certifications:
              type: array
              items:
                $ref: '#/components/schemas/CertificationResponse'
            performanceSummary:
              $ref: '#/components/schemas/PerformanceSummaryResponse'
    ContactResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        role:
          type: string
        email:
          type: string
        phone:
          type: string
        isPrimary:
          type: boolean
        createdAt:
          type: string
          format: date-time
    MaterialCategoryResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
    CapabilityResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        isActive:
          type: boolean
    CertificationResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        documentType:
          $ref: '#/components/schemas/CertificationType'
        documentName:
          type: string
        issueDate:
          type: string
          format: date
        expirationDate:
          type: string
          format: date
        externalFileRef:
          type: string
        isExpired:
          type: boolean
        isExpiringSoon:
          type: boolean
          description: Expiring within 30 days
        createdAt:
          type: string
          format: date-time
...
    PerformanceSummaryResponse:
      type: object
      properties:
        averageRating:
          type: number
          format: float
        evaluationCount:
          type: integer
        lastEvaluationDate:
          type: string
          format: date
        ratingsByCategory:
          type: object
          additionalProperties:
            type: number
            format: float
...
    SupplierValidationResponse:
      type: object
      properties:
        supplierId:
          type: string
          format: uuid
        companyName:
          type: string
        status:
          $ref: '#/components/schemas/SupplierStatus'
        isValid:
          type: boolean
    SupplierEligibilityResponse:
      type: object
      properties:
        supplierId:
          type: string
          format: uuid
        companyName:
          type: string
        isEligible:
          type: boolean
        reasons:
          type: array
          items:
            type: string
          description: Reasons for ineligibility if not eligible
        status:
          $ref: '#/components/schemas/SupplierStatus'
        hasValidCertifications:
          type: boolean
        expiringCertificationsCount:
          type: integer

    # Error Responses
    ErrorResponse:
      type: object
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        traceId:
          type: string
    ValidationErrorResponse:
      allOf:
        - $ref: '#/components/schemas/ErrorResponse'
        - type: object
          properties:
            errors:
              type: object
              additionalProperties:
                type: array
                items:
                  type: string
    DependencyErrorResponse:
      allOf:
        - $ref: '#/components/schemas/ErrorResponse'
        - type: object
          properties:
            dependencies:
              type: array
              items:
                type: object
                properties:
                  serviceName:
                    type: string
                  hasReferences:
                    type: boolean
                  referenceCount:
                    type: integer